0.8.9.0
Implemented reactive armor item
Implemented nanofuel item

Bug fixed: in inventory menu infuse button is enabled from the start with no items to infuse added

+++ Assets/Scripts/Gameplay/Loot/LootConfig.cs
@@ -50,8 +50,8 @@ namespace SpaceAce.Gameplay.Loot
 
         private int CompareItemConfigsByRarityDescending(InventoryItemConfig c1, InventoryItemConfig c2)
         {
-            if ((int)c1.Rarity < (int)c2.Rarity) return 1;
-            if ((int)c1.Rarity > (int)c2.Rarity) return -1;
+            if (c1.Rarity < c2.Rarity) return 1;
+            if (c1.Rarity > c2.Rarity) return -1;
 
             return 0;
         }
